+++ tests/testthat/test-eco_guess.R
@@ -11,8 +11,11 @@ test_that("eco_guess works", {
   # Test that eco guess guesses duplicates
   expect_equal(eco_guess(c("pinus elderica", "pinus elderica"), "common"), c("Afghan pine", "Afghan pine"))
 
+  # Test that warning message pops up when it should
+  expect_warning(val <- eco_guess(as.factor("Common fig"), "botanical"))
+
   # Test that eco guess converts factor to character
-  expect_equal(is.character(eco_guess(as.factor("Common fig"), "botanical")), TRUE)
+  expect_true(is.character(val), TRUE)
 
   # Test that eco guess throws an error if guess arg isn't 'common' or 'botanical'
   expect_error(eco_guess("common fig", "botancal"), "Guess arg isn't 'common' or 'botanical'.")